+++ seek.automation.stub.tests/UsageTests/FromPactBrokerTests.cs
@@ -15,7 +15,7 @@ namespace seek.automation.stub.tests.UsageTests
         {
             _pactAsString = File.ReadAllText("Data/SimplePact.json");
         }
-
+        
         [Fact]
         public void Validate_When_Request_Is_Matched()
         {
@@ -46,7 +46,7 @@ namespace seek.automation.stub.tests.UsageTests
             fakePactBroker.Dispose();
 
             response.StatusCode.ToString().Should().Be("551");
-            response.StatusDescription.Should().Be("Stub on port 9000 says interaction not found. Please verify that the pact associated with this port contains the following request(case insensitive) : Method 'POST', Path '/please/give/me/some/food', Body ''");
+            response.StatusDescription.Should().Be("Stub on port 9000 says interaction not found. Please verify that the pact associated with this port contains the following request(case insensitive) : Method 'POST', Path '/please/give/me/some/food', Body ''. If you have specified filters please also check them.");
         }
 
         [Fact]