+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+### New in 0.0.9 (Released 2016/07/14)
+* Enhanced to add the ability to filter based on the provider_state and description properties of an interaction
+* Enhanced to allow clearing of the filters
+
 ### New in 0.0.8 (Released 2016/07/04)
 * Update Serilog to the latest